Change `DB.query` to support named parameters
This commit changes the `query` method of the `DB` class to
support named parameters.

Instead of relying on the rest parameter (`...values`), a second
parameter is added to `query`, which should be an object mapping
each named parameter to a value:

```
db.query(
  "INSERT INTO tbl (a) value (:val)",
  { val: 123 }
)
```

This is a breaking change as the old array-based parameters method
is not supported anymore.

More info about named parameters is provided in the SQLite docs:
https://tool.oschina.net/uploads/apidocs/sqlite/c3ref/bind_blob.html