Return Sqlite status code on error
When an error occurs, the library will throw an exception.
However, The Error object that is thrown does not contain the actual status code
from Sqlite. This commit addresses this by extending the Error class to
be able to contain an error code member property.

All Sqlite error codes are provided in constants.js now.
https://www.sqlite.org/c3ref/c_abort.html